Profiles-Rest-API/requirements.txt
Django==1.11
djangorestframework==3.6.2
pytz==2019.3
uWSGI==2.0.15
